server/ai_groq_service/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
pillow
pytesseract
PyPDF2
python-multipart